edit test orders and updated with new test_utils module
