+++ tests/test_users.py
@@ -5,6 +5,7 @@ Test suite for User(s) resources.
 from app import app
 from base64 import b64encode
 from models import User
+from tests.test_utils import open_with_auth, add_user
 from peewee import SqliteDatabase
 from http.client import (OK, NOT_FOUND, NO_CONTENT, BAD_REQUEST,
                          CREATED, CONFLICT, UNAUTHORIZED)
@@ -20,21 +21,6 @@ TEST_DB = SqliteDatabase(':memory:')
 TEST_USER_PSW = 'my_password123@'
 
 
-def _add_user(email=None):
-    """
-    Create a single user in the test database.
-    If an email is provided it will be used, otherwise it will be generated
-    by the function before adding the User to the database.
-    """
-    email = email or 'johndoe{}@email.com'.format(int(random.random() * 100))
-
-    return User.create(
-        first_name='John',
-        last_name='Doe',
-        email=email,
-        password=User.hash_password(TEST_USER_PSW),
-        user_id=uuid.uuid4()
-    )
 
 
 class Testuser:
@@ -53,18 +39,6 @@ class Testuser:
     def setup_method(self, test_method):
         User.delete().execute()
 
-    def open_with_auth(self, url, method, username, password):
-        """Generic call to app for http request. """
-
-        AUTH_TYPE = 'Basic'
-        bytes_auth = bytes('{}:{}'.format(username, password), 'ascii')
-        auth_str = '{} {}'.format(
-            AUTH_TYPE, b64encode(bytes_auth).decode('ascii'))
-
-        return self.app.open(url,
-                             method=method,
-                             headers={'Authorization': auth_str})
-
     def test_get_empty_list__success(self):
         resp = self.app.get(API_ENDPOINT.format('users/'))
 
@@ -73,8 +47,8 @@ class Testuser:
         assert User.select().count() == 0
 
     def test_get_users_list__success(self):
-        user1 = _add_user()
-        user2 = _add_user()
+        user1 = add_user(None, TEST_USER_PSW)
+        user2 = add_user(None,TEST_USER_PSW)
 
         resp = self.app.get(API_ENDPOINT.format('users/'))
 
@@ -117,7 +91,7 @@ class Testuser:
         assert resp.status_code == BAD_REQUEST
 
     def test_post_new_user_email_exists__fail(self):
-        _add_user('mail@gmail.com')
+        add_user('mail@gmail.com', TEST_USER_PSW)
         user = {
             'first_name': 'Mario',
             'last_name': 'Rossi',
@@ -163,31 +137,30 @@ class Testuser:
     def test_delete_user__success(self):
         # TODO: refactor for auth implementation
         email = 'mail@email.it'
-        user = _add_user(email)
+        add_user(email, TEST_USER_PSW)
 
-        user_path = 'users/{}'.format(user.user_id)
-        resp = self.open_with_auth(API_ENDPOINT.format(user_path), 'DELETE',
+        user_path = 'users/{}'.format(email)
+        resp = open_with_auth(self.app, API_ENDPOINT.format(user_path), 'DELETE',
                                    email, TEST_USER_PSW)
-
         assert resp.status_code == NO_CONTENT
         assert User.select().count() == 0
 
     def test_delete_user_dont_exists__fail(self):
-        user = _add_user()
+        user = add_user(None, TEST_USER_PSW)
         wrong_uuid = uuid.UUID(int=user.user_id.int + 1)
         user_path = 'users/{}'.format(wrong_uuid)
-        resp = self.open_with_auth(API_ENDPOINT.format(user_path), 'DELETE',
+        resp = open_with_auth(self.app, API_ENDPOINT.format(user_path), 'DELETE',
                                    user.email, TEST_USER_PSW)
 
         assert resp.status_code == NOT_FOUND
         assert User.select().count() == 1
 
     def test_delete_user_other_user__fail(self):
-        user_A = _add_user('user.a@users.com')
-        user_B = _add_user('user.b@users.com')
+        user_A = add_user('user.a@users.com', TEST_USER_PSW)
+        user_B = add_user('user.b@users.com', TEST_USER_PSW)
 
         path = 'users/{}'.format(user_A.user_id)
-        resp = self.open_with_auth(API_ENDPOINT.format(path), 'DELETE',
+        resp = open_with_auth(self.app, API_ENDPOINT.format(path), 'DELETE',
                                    user_B.user_id, TEST_USER_PSW)
 
         assert resp.status_code == UNAUTHORIZED
@@ -196,10 +169,11 @@ class Testuser:
         assert User.select().count() == 2
 
     def test_delete_user_auth_does_not_exists__fail(self):
-        user = _add_user()
+        user = add_user(None, TEST_USER_PSW)
+
 
         path = 'users/{}'.format(user.user_id)
-        resp = self.open_with_auth(API_ENDPOINT.format(path), 'DELETE',
+        resp = open_with_auth(self.app, API_ENDPOINT.format(path), 'DELETE',
                                    'donot@exists.com', 'unused_psw')
 
         assert resp.status_code == UNAUTHORIZED
